+++ src/main.js
@@ -17,6 +17,7 @@ import fetch from 'node-fetch'
 
 import config from './config'
 import { toJSON } from './utils'
+import { readTelegramCache } from './telegram'
 
 export default () => {
   const
@@ -34,6 +35,7 @@ export default () => {
 
   config
     .map(readCache)
+    .map(readTelegramCache)
     .map(cfg => {
 
       // Fetch JSON string from URL